Single type per file, file name matches class name
