+++ sdk/appconfiguration/Azure.Data.AppConfiguration/src/ConfigurationServiceSerializer.cs
