+++ sdk/core/Azure.Core/src/AzureExtensions.cs
