+++ sdk/core/Azure.Core/src/Buffers/AzureBaseBuffersExtensions.cs
