+++ sdk/core/Azure.Core/src/Http/RequestUriBuilder.cs
