+++ sdk/core/Azure.Core/src/Operation{T}.cs
