+++ sdk/core/Azure.Core/src/Pipeline/ClientDiagnostics.cs
