+++ sdk/core/Azure.Core/src/Pipeline/ClientOptions.cs
