+++ sdk/core/Azure.Core/src/Pipeline/DiagnosticScope.cs
