+++ sdk/core/Azure.Core/src/RequestFailedException.cs
