+++ sdk/core/Azure.Core/src/Shared/NullableAttributes.cs
@@ -3,6 +3,9 @@
 
 #define INTERNAL_NULLABLE_ATTRIBUTES
 
+#pragma warning disable SA1402 // File may only contain a single type
+#pragma warning disable SA1649 // File name should match first type name
+
 namespace System.Diagnostics.CodeAnalysis
 {
     /// <summary>Specifies that null is allowed as an input even if the corresponding type disallows it.</summary>
