+++ sdk/core/Azure.Core/tests/ClientTestBaseDiagnosticScopeTests.cs
@@ -51,67 +51,67 @@ namespace Azure.Core.Tests
             InvalidDiagnosticScopeTestClient client = InstrumentClient(new InvalidDiagnosticScopeTestClient());
             await client.CorrectScopeAsync();
         }
-    }
 
-    public class InvalidDiagnosticScopeTestClient
-    {
-        private void FireScope(string method)
+        public class InvalidDiagnosticScopeTestClient
         {
-            ClientDiagnostics clientDiagnostics = new ClientDiagnostics(true);
-            string activityName = $"{typeof(InvalidDiagnosticScopeTestClient).FullName}.{method}";
-            DiagnosticScope scope = clientDiagnostics.CreateScope(activityName);
-            scope.Start();
-            scope.Dispose();
-        }
+            private void FireScope(string method)
+            {
+                ClientDiagnostics clientDiagnostics = new ClientDiagnostics(true);
+                string activityName = $"{typeof(InvalidDiagnosticScopeTestClient).FullName}.{method}";
+                DiagnosticScope scope = clientDiagnostics.CreateScope(activityName);
+                scope.Start();
+                scope.Dispose();
+            }
 
-        [ForwardsClientCalls]
-        public virtual Task<bool> NoScopeAsync()
-        {
-            return Task.FromResult(true);
-        }
+            [ForwardsClientCalls]
+            public virtual Task<bool> NoScopeAsync()
+            {
+                return Task.FromResult(true);
+            }
 
-        [ForwardsClientCalls]
-        public virtual bool NoScope()
-        {
-            return true;
-        }
+            [ForwardsClientCalls]
+            public virtual bool NoScope()
+            {
+                return true;
+            }
 
-        public virtual Task<bool> WrongScopeAsync()
-        {
-            FireScope("DoesNotExist");
-            return Task.FromResult(true);
-        }
+            public virtual Task<bool> WrongScopeAsync()
+            {
+                FireScope("DoesNotExist");
+                return Task.FromResult(true);
+            }
 
-        public virtual bool WrongScope()
-        {
-            FireScope("DoesNotExist");
-            return true;
-        }
+            public virtual bool WrongScope()
+            {
+                FireScope("DoesNotExist");
+                return true;
+            }
 
-        public virtual Task<bool> CorrectScopeAsync()
-        {
-            FireScope(nameof(CorrectScope));
-            return Task.FromResult(true);
-        }
+            public virtual Task<bool> CorrectScopeAsync()
+            {
+                FireScope(nameof(CorrectScope));
+                return Task.FromResult(true);
+            }
 
-        public virtual bool CorrectScope()
-        {
-            FireScope(nameof(CorrectScope));
-            return true;
-        }
+            public virtual bool CorrectScope()
+            {
+                FireScope(nameof(CorrectScope));
+                return true;
+            }
 
-        [ForwardsClientCalls]
-        public virtual Task<bool> ForwardsAsync()
-        {
-            FireScope(nameof(CorrectScope));
-            return Task.FromResult(true);
-        }
+            [ForwardsClientCalls]
+            public virtual Task<bool> ForwardsAsync()
+            {
+                FireScope(nameof(CorrectScope));
+                return Task.FromResult(true);
+            }
 
-        [ForwardsClientCalls]
-        public virtual bool Forwards()
-        {
-            FireScope(nameof(CorrectScope));
-            return true;
+            [ForwardsClientCalls]
+            public virtual bool Forwards()
+            {
+                FireScope(nameof(CorrectScope));
+                return true;
+            }
         }
     }
 }
