+++ sdk/core/Azure.Core/tests/RecordSessionTests.cs
