+++ sdk/core/Azure.Core/tests/TestEventListener.cs
@@ -40,10 +40,4 @@ namespace Azure.Core.Testing
             base.Dispose();
         }
     }
-
-    public static class TestEventListenerExtensions
-    {
-        public static T GetProperty<T>(this EventWrittenEventArgs data, string propName)
-            => (T)data.Payload[data.PayloadNames.IndexOf(propName)];
-    }
 }
