+++ sdk/core/Azure.Core/tests/TestEventListenerExtensions.cs
@@ -1,0 +1,13 @@
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+using System.Diagnostics.Tracing;
+
+namespace Azure.Core.Testing
+{
+    internal static class TestEventListenerExtensions
+    {
+        public static T GetProperty<T>(this EventWrittenEventArgs data, string propName)
+            => (T)data.Payload[data.PayloadNames.IndexOf(propName)];
+    }
+}
