+++ sdk/core/Azure.Core/tests/TestFramework/RandomExtensions.cs
@@ -1,0 +1,17 @@
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+using System;
+
+namespace Azure.Core.Testing
+{
+    public static class RandomExtensions
+    {
+        public static Guid NewGuid(this Random random)
+        {
+            var bytes = new byte[16];
+            random.NextBytes(bytes);
+            return new Guid(bytes);
+        }
+    }
+}
