+++ sdk/core/Microsoft.Extensions.Azure/src/Internal/AzureClientCredentialOptions{TClient}.cs
