+++ sdk/core/Microsoft.Extensions.Azure/src/Internal/ClientFactory.cs
