+++ sdk/core/Microsoft.Extensions.Azure/src/Internal/ClientRegistration{TClient,TOptions}.cs
