+++ sdk/identity/Azure.Identity/tests/Mock/MockManagedIdentityClient.cs
@@ -90,86 +90,4 @@ namespace Azure.Identity.Tests.Mock
             return _tokenFactory(scopes, tenantId, clientId, clientSecret, cancellationToken);
         }
     }
-
-    internal class MockAadClient : AadIdentityClient
-    {
-        private static AccessToken ExpiredTokenFactory(string[] scopes, string tenantId, string clientId, string clientSecret, CancellationToken cancellationToken)
-        {
-            return CreateAccessToken(scopes, tenantId, clientId, clientSecret, DateTimeOffset.UtcNow - TimeSpan.FromMinutes(1));
-        }
-
-        private static AccessToken LiveTokenFactory(string[] scopes, string tenantId, string clientId, string clientSecret, CancellationToken cancellationToken)
-        {
-            return CreateAccessToken(scopes, tenantId, clientId, clientSecret, DateTimeOffset.UtcNow + TimeSpan.FromHours(1));
-        }
-
-        private static AccessToken CreateAccessToken(string[] scopes, string tenantId, string clientId, string clientSecret, DateTimeOffset expires)
-        {
-            MockToken token = CreateMockToken(scopes, tenantId, clientId, clientSecret);
-
-            return new AccessToken(token.ToString(), expires);
-        }
-
-        private static MockToken CreateMockToken(string[] scopes, string tenantId, string clientId, string clientSecret)
-        {
-            return new MockToken().WithField("scopes", string.Join("+", scopes)).WithField("tenantId", tenantId).WithField("clientId", clientId).WithField("clientSecret", clientSecret);
-        }
-
-        public static MockAadClient ExpiredTokenClient { get; } = new MockAadClient(ExpiredTokenFactory);
-
-        public static MockAadClient LiveTokenClient { get; } = new MockAadClient(LiveTokenFactory);
-
-        private readonly Func<string[], string, string, string, CancellationToken, AccessToken> _tokenFactory;
-
-        public MockAadClient()
-            : this(LiveTokenFactory)
-        {
-
-        }
-
-        public MockAadClient(AccessToken token)
-            : this(() => token)
-        {
-        }
-
-        public MockAadClient(Func<AccessToken> tokenFactory)
-            : this((scopes, tenantId, clientId, clientSecret, cancellationToken) => tokenFactory())
-        {
-        }
-
-        public MockAadClient(Func<string[], string, string, string, CancellationToken, AccessToken> tokenFactory)
-        {
-            _tokenFactory = tokenFactory;
-        }
-
-        public override AccessToken Authenticate(string tenantId, string clientId, string clientSecret, string[] scopes, CancellationToken cancellationToken = default)
-        {
-            return CreateToken(scopes, clientId: clientId, tenantId: tenantId, clientSecret: clientSecret, cancellationToken: cancellationToken);
-        }
-
-        public override async Task<AccessToken> AuthenticateAsync(string tenantId, string clientId, string clientSecret, string[] scopes, CancellationToken cancellationToken = default)
-        {
-            return await CreateTokenAsync(scopes, clientId: clientId, tenantId: tenantId, clientSecret: clientSecret, cancellationToken: cancellationToken);
-        }
-
-        private async Task<AccessToken> CreateTokenAsync(string[] scopes, string tenantId = default, string clientId = default, string clientSecret = default, CancellationToken cancellationToken = default)
-        {
-            if (cancellationToken != default)
-            {
-                await Task.Delay(1000, cancellationToken);
-            }
-
-            return _tokenFactory(scopes, tenantId, clientId, clientSecret, cancellationToken);
-        }
-
-        private AccessToken CreateToken(string[] scopes, string tenantId = default, string clientId = default, string clientSecret = default, CancellationToken cancellationToken = default)
-        {
-            if (cancellationToken != default)
-            {
-                Task.Delay(1000, cancellationToken).GetAwaiter().GetResult();
-            }
-
-            return _tokenFactory(scopes, tenantId, clientId, clientSecret, cancellationToken);
-        }
-    }
 }
